Created links stories file.
